requirements/base.txt
Django==4.1.7
PyJWT==2.6.0
